rename index modules to find them easier
