+++ packages/chealt.com/public/Documents/Documents.js
@@ -1,18 +1,18 @@
 import { useCallback, useState, useRef } from 'preact/hooks';
 
-import EmptyState from '../EmptyState';
+import EmptyState from '../EmptyState/EmptyState';
 import Button from '../Form/Button';
 import FileInput from '../Form/FileInput';
 import { toggleItem } from '../Helpers/array';
 import DocumentsIcon from '../Icons/Documents';
 import { useObjectStore } from '../IndexedDB/hooks';
-import PageTitle from '../PageTitle';
-import { add as addToast } from '../Toast';
+import PageTitle from '../PageTitle/PageTitle';
+import { add as addToast } from '../Toast/Toast';
 import Controls from './Controls';
 import Item from './Item';
 import { getFilesFromEvent } from './utils';
 
-import styles from './index.module.css';
+import styles from './Documents.module.css';
 
 const bySavedTime = (a, b) => b.value.savedTimestamp - a.value.savedTimestamp;
