+++ packages/chealt.com/public/EmptyState/EmptyState.js
@@ -1,4 +1,4 @@
-import styles from './index.module.css';
+import styles from './EmptyState.module.css';
 
 const EmptyState = ({ children }) => <div class={styles.emptyState}>{children}</div>;
