+++ packages/chealt.com/public/Profiles/Profiles.js
@@ -5,8 +5,8 @@ import { useObjectStore } from '../IndexedDB/hooks';
 import Container from '../Layout/Container';
 import List from '../List/List';
 import ListItem from '../List/ListItem';
-import Modal from '../Modal';
-import PageTitle from '../PageTitle';
+import Modal from '../Modal/Modal';
+import PageTitle from '../PageTitle/PageTitle';
 import ProfileForm from './ProfileForm';
 import { setSelectedProfileId } from './signals';
