+++ packages/chealt.com/public/QRCode/QRCode.js
