+++ packages/chealt.com/public/Table/Table.js
@@ -1,4 +1,4 @@
-import styles from './index.module.css';
+import styles from './Table.module.css';
 
 const Table = ({ children }) => <table class={styles.table}>{children}</table>;
