+++ packages/chealt.com/public/Vaccinations/NewItem.js
@@ -3,7 +3,7 @@ import { useState } from 'preact/hooks';
 import Button from '../Form/Button';
 import Form from '../Form/Form';
 import Input from '../Form/Input';
-import { add as addToast } from '../Toast';
+import { add as addToast } from '../Toast/Toast';
 import { getConditions, getLocalVaccinations } from './utils';
 
 import styles from './NewItem.module.css';
