+++ packages/chealt.com/public/Vaccinations/Vaccinations.js
@@ -1,20 +1,20 @@
 import { localFormatDate } from '@chealt/browser-utils';
 import { useCallback, useState } from 'preact/hooks';
 
-import EmptyState from '../EmptyState';
+import EmptyState from '../EmptyState/EmptyState';
 import Button from '../Form/Button';
 import Input from '../Form/Input';
 import Tag from '../Form/Tag';
 import { toggleItem } from '../Helpers/array';
 import Vaccine from '../Icons/Vaccine';
 import { useObjectStore } from '../IndexedDB/hooks';
-import Modal from '../Modal';
-import PageTitle from '../PageTitle';
-import Table from '../Table';
+import Modal from '../Modal/Modal';
+import PageTitle from '../PageTitle/PageTitle';
 import Cell from '../Table/Cell';
 import HeadCell from '../Table/HeadCell';
 import Row from '../Table/Row';
-import { add as addToast } from '../Toast';
+import Table from '../Table/Table';
+import { add as addToast } from '../Toast/Toast';
 import Controls from './Controls';
 import NewItem from './NewItem';
