+++ packages/chealt.com/public/pages/Home.js
@@ -1,1 +1,1 @@
-export { default } from '../Home';
+export { default } from '../Home/Home';
