+++ packages/chealt.com/public/pages/Share.js
@@ -1,1 +1,1 @@
-export { default } from '../Share';
+export { default } from '../Share/Share';
